update wait message when generating masternode key
When generating a masternode key, if the wallet is not yet loaded, the script will wait 30 seconds and try again. The wait message was a little unclear if user needed to do anything or not, so changed it to be clear that the user should not do anything while wallet waits 30 seconds and tries again.